+++ shared/types/bridge.ts
@@ -1,10 +1,6 @@
+import { BigNumber } from 'bignumber.js';
 import { Hex } from '@metamask/utils';
 import { SwapsTokenObject } from '../constants/swaps';
-import {
-  L1GasFees,
-  QuoteRequest,
-  QuoteResponse,
-} from '../../ui/pages/bridge/types';
 import { RequestStatus } from '../../app/scripts/controllers/bridge/constants';
 
 export type ChainConfiguration = {
@@ -55,3 +51,145 @@ export enum BridgeBackgroundAction {
   RESET_STATE = 'resetState',
   GET_BRIDGE_ERC20_ALLOWANCE = 'getBridgeERC20Allowance',
 }
+
+export type L1GasFees = {
+  l1GasFeesInHexWei?: string; // l1 fees for approval and trade in hex wei, appended by controller
+};
+
+// Values derived from the quote response
+// valueInCurrency values are calculated based on the user's selected currency
+export type QuoteMetadata = {
+  gasFee: { amount: BigNumber; valueInCurrency: BigNumber | null };
+  totalNetworkFee: { amount: BigNumber; valueInCurrency: BigNumber | null }; // estimatedGasFees + relayerFees
+  totalMaxNetworkFee: { amount: BigNumber; valueInCurrency: BigNumber | null }; // maxGasFees + relayerFees
+  adjustedReturn: { valueInCurrency: BigNumber | null }; // destTokenAmount - totalNetworkFee
+  sentAmount: { amount: BigNumber; valueInCurrency: BigNumber | null }; // srcTokenAmount + metabridgeFee
+  swapRate: BigNumber; // destTokenAmount / sentAmount
+  cost: { valueInCurrency: BigNumber | null }; // sentAmount - adjustedReturn
+};
+
+// Sort order set by the user
+export enum SortOrder {
+  COST_ASC = 'cost_ascending',
+  ETA_ASC = 'time_descending',
+}
+
+// Types copied from Metabridge API
+export enum BridgeFlag {
+  EXTENSION_CONFIG = 'extension-config',
+}
+
+type DecimalChainId = string;
+export type GasMultiplierByChainId = Record<DecimalChainId, number>;
+
+export type FeatureFlagResponse = {
+  [BridgeFlag.EXTENSION_CONFIG]: {
+    refreshRate: number;
+    maxRefreshCount: number;
+    support: boolean;
+    chains: Record<number, ChainConfiguration>;
+  };
+};
+
+export type BridgeAsset = {
+  chainId: ChainId;
+  address: string;
+  symbol: string;
+  name: string;
+  decimals: number;
+  icon?: string;
+};
+
+export type QuoteRequest = {
+  walletAddress: string;
+  destWalletAddress?: string;
+  srcChainId: ChainId;
+  destChainId: ChainId;
+  srcTokenAddress: string;
+  destTokenAddress: string;
+  srcTokenAmount: string; // This is the amount sent
+  slippage: number;
+  aggIds?: string[];
+  bridgeIds?: string[];
+  insufficientBal?: boolean;
+  resetApproval?: boolean;
+  refuel?: boolean;
+};
+
+type Protocol = {
+  name: string;
+  displayName?: string;
+  icon?: string;
+};
+
+enum ActionTypes {
+  BRIDGE = 'bridge',
+  SWAP = 'swap',
+  REFUEL = 'refuel',
+}
+
+type Step = {
+  action: ActionTypes;
+  srcChainId: ChainId;
+  destChainId?: ChainId;
+  srcAsset: BridgeAsset;
+  destAsset: BridgeAsset;
+  srcAmount: string;
+  destAmount: string;
+  protocol: Protocol;
+};
+
+type RefuelData = Step;
+
+export type Quote = {
+  requestId: string;
+  srcChainId: ChainId;
+  srcAsset: BridgeAsset;
+  // Some tokens have a fee of 0, so sometimes it's equal to amount sent
+  srcTokenAmount: string; // Atomic amount, the amount sent - fees
+  destChainId: ChainId;
+  destAsset: BridgeAsset;
+  destTokenAmount: string; // Atomic amount, the amount received
+  feeData: Record<FeeType.METABRIDGE, FeeData> &
+    Partial<Record<FeeType, FeeData>>;
+  bridgeId: string;
+  bridges: string[];
+  steps: Step[];
+  refuel?: RefuelData;
+};
+
+export type QuoteResponse = {
+  quote: Quote;
+  approval: TxData | null;
+  trade: TxData;
+  estimatedProcessingTimeInSeconds: number;
+};
+
+export enum ChainId {
+  ETH = 1,
+  OPTIMISM = 10,
+  BSC = 56,
+  POLYGON = 137,
+  ZKSYNC = 324,
+  BASE = 8453,
+  ARBITRUM = 42161,
+  AVALANCHE = 43114,
+  LINEA = 59144,
+}
+
+export enum FeeType {
+  METABRIDGE = 'metabridge',
+  REFUEL = 'refuel',
+}
+export type FeeData = {
+  amount: string;
+  asset: BridgeAsset;
+};
+export type TxData = {
+  chainId: ChainId;
+  to: string;
+  from: string;
+  value: string;
+  data: string;
+  gasLimit: number | null;
+};
